OrderGroup CRUD

+++ src/main/java/com/example/study/model/entity/Item.java
@@ -10,8 +10,10 @@ import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.function.BiFunction;
 
 @Data
 @AllArgsConstructor
@@ -35,7 +37,7 @@ public class Item {
 
     private String content;
 
-    private Integer price;
+    private BigDecimal price;
 
     private String brandName;
